Phase 3: x86_64 userspace (syscall/sysret + 64-bit crt)
Kernel side:
- syscall/sysret path with EFER.SCE, STAR/LSTAR/SFMASK; syscall_entry
  saves the full registers_t (syscall number preserved), and the return
  path restores all user GPRs before sysretq.
- arch_setup_first_stack fork frames now include the GPR block, pushed in
  reverse pop order so RAX=0 lands where first_entry_trampoline pops it
  (fork() returns 0 in the child); first_entry_trampoline pops 15 GPRs.
- Boot page tables are accessed via the direct map (PHYS_TO_VIRT) so they
  work under a user PML4; arch_mm_free_user_tables keeps the PML4 for exec
  reuse and a new arch_mm_free_pgd() frees it at teardown.
- PIC fallback in interrupt_init re-enables already-registered IRQs
  (fixes a silent timer IRQ0 mask when APIC is unavailable).
- Fixed 32-bit pointer truncations in VIRT_TO_PHYS() callers (pagefault,
  fork, exec, spawn) and free_task's page_free.
- kbd/char/tty/irq/timer debug instrumentation removed; a single
  [EXEC] pid N: name log line remains on successful exec.

Userspace (crt):
- x86_64 build (crt/build64): -m64, crt0_64.s (SysV, parses the initial
  stack, aligns RSP), linker64.ld (ELF64 flat at 0x08000000).
- syscall wrappers use the 'syscall' instruction (RDI/RSI/RDX/R10/R8/R9);
  i386 keeps int 0x80 via __x86_64__ guard.
- time/utime/fb ABI structs pinned to 32-bit fields (x32-compatible);
  init passes environ to execve (fork clobbers the envp register).

Boots under qemu-system-x86_64 + OVMF + EFI GRUB: init mounts /dev,/proc,
/tmp, spawns sh, and typed commands fork+exec (verified 'ls': pid 3).
i386 regression boot verified.

+++ crt/include/sys/fb.h
@@ -57,17 +57,17 @@ struct fb_var_screeninfo {
 /* Fixed screen information (Linux 32-bit layout, 64 bytes). */
 struct fb_fix_screeninfo {
     char             id[16];
-    unsigned long    smem_start;
-    unsigned long    smem_len;
+    uint32_t        smem_start;
+    uint32_t        smem_len;
     unsigned int     type;
     unsigned int     type_aux;
     unsigned int     visual;
     unsigned short   xpanstep;
     unsigned short   ypanstep;
     unsigned short   ywrapstep;
-    unsigned long    line_length;
-    unsigned long    mmio_start;
-    unsigned long    mmio_len;
+    uint32_t        line_length;
+    uint32_t        mmio_start;
+    uint32_t        mmio_len;
     unsigned int     accel;
     unsigned short   capabilities;
     unsigned short   reserved[2];
